Added support to use lists as the sides of a die. Changed keywords to all-lowercase. Added support to operations between singular values and lists. Moved from 64 bit numerals to 32 bit. Fixed function variable resolution.

+++ src/common_util.rs
@@ -2,8 +2,8 @@ use std::fmt;
 use std::collections::HashMap;
 use std::hash::Hash;
 
-pub type IntType = i64;
-pub type FloatType = f64;
+pub type IntType = i32;
+pub type FloatType = f32;
 
 pub trait Pow<RHS = Self> {
 	type Output;
